refactor: split imposter/predicates.rs into a predicates/ module
Pure code-movement, no logic change. The 69KB single file becomes:
- predicates/mod.rs: stub_matches, parse_query, predicate_matches dispatch,
  parse_query_string, and the inline test module
- predicates/fields.rs: check_predicate_fields / check_predicate_fields_regex
- predicates/json.rs: json_value_to_string, check_exists_json_recursive,
  check_exists_predicate, compare_json_recursive

Cross-module helpers are pub(crate); the public paths (predicates::stub_matches,
predicate_matches, parse_query_string, parse_query) are unchanged, so external
imports and re-exports still resolve.
